requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0

# Weaviate Vector Database
weaviate-client==4.4.1
numpy>=1.21.0
pandas>=1.3.0

# FastAPI (for API endpoints)
fastapi>=0.104.0
uvicorn>=0.24.0

# Logging and utilities
rich>=13.0.0
pydantic>=2.0.0
